test: count the shim's spawn attempts as size_t

+++ test/bundler/bun-build-thread-spawn-failure.test.ts
@@ -30,7 +30,7 @@ const SHIM_C = /* c */ `
 static int (*real_pthread_create)(pthread_t *, const pthread_attr_t *, void *(*)(void *), void *);
 static const char *plan;
 static size_t plan_len;
-static int attempts;
+static size_t attempts;
 
 __attribute__((constructor)) static void init(void) {
   real_pthread_create = dlsym(RTLD_NEXT, "pthread_create");
